Successes : Got the profile info data to update on the front end. Backend is setup. Need to create queries in Django.
Ditched Attempt: Ditched the attempt to use createContext and useContext. Did more research and this is not the best approach as it makes your variables global and rerenders all of the effected components. Costly time. Try sticking to useState and useEffect for this project for simplicity.

Bug fix: Deleted id from all the form controls because id takes precedence over control id.

TODO: Place hobbies selections into backend only? Or keep in front end and just send to backend? Discuss with Jacob.