Major push: Added friend request functionality.
Note: Permissions only allow for the sender to delete the request and the receiver to modify the request status. Updating the status to Accept or Decline will delete the request with one key difference, Accept saves the accepted friend to the receivers friends field. It's many-to-many, so this will reciprocate the friend relationship in the other users friend's list.

Todo: Make sure users can't make dual requests to each other. .Once a request is in, then the other can't request.